LayoutTests/fast/js/sputnik/Conformance/12_Statement/12.6_Iteration_Statements/12.6.1_The_do_while_Statement/S12.6.1_A15-expected.txt

CONSOLE MESSAGE: line 80: SyntaxError: Parse error
S12.6.1_A15

PASS Expected parsing failure

TEST COMPLETE
